Do not attempt to build Propeller on unsupported host OSes
The build tools for Propeller are supplied as binary executables
for the Linux PC environment, and MacOS (64-bit only),

Currently the build attempts to build the Propeller files on
Linux but does not check the CPU architecture of the host OS.

A build failure was observed when trying to do a build using
a Raspberry Pi4 running 64-bit RaspberryPiOS (buster). The failure
was due to attempting to run x86 ELF files of the Propeller
build tools on a Aarch64 (ARMv8-A) RPi4 architecture which
obviously will fail to run.

It is noted that the build tools are prepared early in the build
process despite the Propeller tools not being needed for many
of the build targets.

This commit resolves the build defect by modifying Source/Makefile
to skip over building the Propeller source files when the host
architecture does not support the binary Propeller build tools.
Only 32-bit and 64-bit x86 Linux host OSes and 64-bit MacOS will
build the Propeller files.